LayoutTests/svg/animations/svginteger-animation-1-expected.txt

SVG 1.1 dynamic animation tests

Test animation of SVGInteger.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS feConvolveMatrix.targetX.animVal is 0
PASS feConvolveMatrix.targetX.baseVal is 0
PASS feConvolveMatrix.targetX.animVal is 1
PASS feConvolveMatrix.targetX.baseVal is 0
PASS feConvolveMatrix.targetX.animVal is 2
PASS feConvolveMatrix.targetX.baseVal is 0
PASS feConvolveMatrix.targetX.animVal is 0
PASS feConvolveMatrix.targetX.baseVal is 0
PASS successfullyParsed is true

TEST COMPLETE
